+++ lizard_progress/admin.py
@@ -24,6 +24,11 @@ class ProjectTypeAdmin(admin.ModelAdmin):
     list_editable = ('name', 'organization', 'default')
 
 
+class ExportRunAdmin(admin.ModelAdmin):
+    model = models.ExportRun
+    search_fields = ['activity', 'exporttype']
+
+
 admin.site.register(models.Hydrovak)
 admin.site.register(models.Location)
 admin.site.register(models.AvailableMeasurementType)
@@ -31,7 +36,14 @@ admin.site.register(models.Project)
 admin.site.register(models.Activity)
 admin.site.register(models.Organization, OrganizationAdmin)
 admin.site.register(models.UserProfile)
+admin.site.register(models.UserRole)
 admin.site.register(models.ErrorMessage)
 admin.site.register(models.LizardConfiguration)
 admin.site.register(models.ProjectType, ProjectTypeAdmin)
-admin.site.register(models.ExportRun)
+admin.site.register(models.ExportRun, ExportRunAdmin)
+admin.site.register(models.ExpectedAttachment)
+admin.site.register(models.MeasurementTypeAllowed)
+admin.site.register(models.Measurement)
+admin.site.register(models.UploadedFile)
+admin.site.register(models.UploadedFileError)
+admin.site.register(models.UploadLog)